Reclassify previously invalidated tests
Reclassify previously invalidated typelevel tests

 - some of them work again
 - some of them can be turned into neg tests
 - the rest gets a note why they are invalid
