+++ tests/invalid/neg/typelevel.scala
@@ -1,3 +1,4 @@
+// This test requires retyping from untyped trees after inlining, which is not supported anymore
 trait HList {
   def length: Int
   def head: Any
